remove the Content_Helpers class
